Add suitable Representation for read-in models
We already create a dummy Representation for new Models
to satisfy python-ihm's sanity checks (e.g. that each
Atom matches an atomic representation segment). But when
reading in a Model from a file, its Assembly is initially
empty, so the Representation is empty too. Correct this
at System-finalize time. Closes #20.