Allow describing custom chemical compounds
Use new functionality in python-ihm to extend
ihm.ChemComp to allow for chemical compounds
that are not defined in wwPDB's CCD but are
in ModelArchive's CCD (add ccd='ma' to the ChemComp
constructor). Also allow for custom chemicals
to be defined in the mmCIF file itself
(with ccd='local'). Add a new modelcif.descriptor
module with a range of suitable chemical descriptors.
Closes #21.