Allow concatenation of associated files
Use new functionality in python-ihm to
add a new argument to modelcif.reader.read(),
which specifies an existing System object.
In this case, any data read from the file
will be added to that System, rather than placed
into new System objects. This is primarily
intended to allow the reading of associated
files; relates #10.