Read dihedral QA metrics
Dihedral QA metrics are keyed by atom_site.id
but we store them in the Model. Make a simple
mapping from atom ID to model num when we
read the atom_site table. We assume that atom
IDs are numeric and roughly sequential (may
contain gaps). Use this mapping to put the
new metric object in the right model list.